Revert "added sprinting"
This reverts commit 61b593cb934770af7f7dcb115eaa8346302d0f0a.

+++ GMD1/Assets/Scripts/PlayerScripts/HealthBarScript.cs
@@ -7,12 +7,13 @@ public class HealthBarScript : MonoBehaviour
 {
 
     public Slider slider;
-    public void setHealth(float health){
+    // Start is called before the first frame update
+    public void setHealth(int health){
 
         slider.value=health;
     }
 
-    public void setMaxHealth(float maxHealth){
+    public void setMaxHealth(int maxHealth){
 
         slider.maxValue=maxHealth;
         slider.value=maxHealth;
